+++ src/client/datascience/notebook/executionHelpers.ts
@@ -7,10 +7,9 @@ import type { nbformat } from '@jupyterlab/coreutils';
 import type { KernelMessage } from '@jupyterlab/services';
 import { NotebookCell, NotebookCellRunState, NotebookDocument } from 'vscode';
 import { createErrorOutput } from '../../../datascience-ui/common/cellFactory';
-import { IDisposable } from '../../common/types';
 import { INotebookModelModifyChange } from '../interactive-common/interactiveWindowTypes';
 import { ICell, INotebookModel } from '../types';
-import { cellOutputsToVSCCellOutputs, findMappedNotebookCellData, translateErrorOutput } from './helpers';
+import { cellOutputsToVSCCellOutputs, translateErrorOutput } from './helpers';
 
 export function hasTransientOutputForAnotherCell(output?: nbformat.IOutput) {
     return (
@@ -76,67 +75,6 @@ export function handleUpdateDisplayDataMessage(
 export function updateCellWithErrorStatus(cell: NotebookCell, ex: Partial<Error>) {
     cell.outputs = [translateErrorOutput(createErrorOutput(ex))];
     cell.metadata.runState = NotebookCellRunState.Error;
-}
-
-/**
- * Responsible for syncing changes from our model into the VS Code cells.
- * Eg. when executing a cell, we update our model with the output, and here we react to those events and update the VS Code output.
- * This way, all updates to VSCode cells can happen in one place (here), and we can focus on updating just the Cell model with the data.
- * Here we only keep the outputs in sync. The assumption is that we won't be adding cells directly.
- * If adding cells and the like then please use VSC api to manipulate cells, else we have 2 ways of doing the same thing and that could lead to issues.
- */
-export function monitorModelCellOutputChangesAndUpdateNotebookDocument(
-    document: NotebookDocument,
-    model: INotebookModel
-): IDisposable {
-    let wasUntitledNotebook = model.isUntitled;
-    let stopSyncingOutput = false;
-    const disposable = model.changed((change) => {
-        if (stopSyncingOutput) {
-            return;
-        }
-        if (change.kind === 'saveAs') {
-            if (wasUntitledNotebook) {
-                wasUntitledNotebook = false;
-                // User saved untitled file as a real file.
-                return;
-            } else {
-                // Ok, user save a normal notebook as another name.
-                // Stop monitoring changes.
-                stopSyncingOutput = true;
-                disposable.dispose();
-                return;
-            }
-        }
-        // We're only interested in updates to cells.
-        if (change.kind !== 'modify') {
-            return;
-        }
-        for (const cell of change.newCells) {
-            const uiCellToUpdate = findMappedNotebookCellData(cell, document.cells);
-            if (!uiCellToUpdate) {
-                continue;
-            }
-            const newOutput = Array.isArray(cell.data.outputs)
-                ? // tslint:disable-next-line: no-any
-                  cellOutputsToVSCCellOutputs(cell.data.outputs as any)
-                : [];
-            // If there were no cells and still no cells, nothing to update.
-            if (newOutput.length === 0 && uiCellToUpdate.outputs.length === 0) {
-                return;
-            }
-            // If no changes in output, then nothing to do.
-            if (
-                newOutput.length === uiCellToUpdate.outputs.length &&
-                JSON.stringify(newOutput) === JSON.stringify(uiCellToUpdate.outputs)
-            ) {
-                return;
-            }
-            uiCellToUpdate.outputs = newOutput;
-        }
-    });
-
-    return disposable;
 }
 
 /**
