+++ src/client/datascience/notebook/integration.ts
@@ -6,11 +6,15 @@ import * as path from 'path';
 import { notebook } from 'vscode';
 import { IExtensionSingleActivationService } from '../../activation/types';
 import { ICommandManager } from '../../common/application/types';
+import { UseProposedApi } from '../../common/constants';
 import { NativeNotebook } from '../../common/experimentGroups';
 import { IFileSystem } from '../../common/platform/types';
 import { IDisposableRegistry, IExperimentsManager, IExtensionContext } from '../../common/types';
 import { noop } from '../../common/utils/misc';
+import { IServiceManager } from '../../ioc/types';
+import { INotebookEditorProvider } from '../types';
 import { NotebookContentProvider } from './contentProvider';
+import { NotebookEditorProvider } from './notebookEditorProvider';
 import { NotebookKernel } from './notebookKernel';
 
 /**
@@ -27,7 +31,9 @@ export class NotebookIntegration implements IExtensionSingleActivationService {
         @inject(IExtensionContext) private readonly context: IExtensionContext,
         @inject(IFileSystem) private readonly fs: IFileSystem,
         @inject(ICommandManager) private readonly commandManager: ICommandManager,
-        @inject(NotebookKernel) private readonly notebookKernel: NotebookKernel
+        @inject(NotebookKernel) private readonly notebookKernel: NotebookKernel,
+        @inject(IServiceManager) private readonly serviceManager: IServiceManager,
+        @inject(UseProposedApi) private readonly useProposedApi: boolean
     ) {}
     public async activate(): Promise<void> {
         // This condition is temporary.
@@ -36,6 +42,20 @@ export class NotebookIntegration implements IExtensionSingleActivationService {
         if (!this.experiment.inExperiment(NativeNotebook.experiment)) {
             return;
         }
+
+        // This condition is temporary.
+        // If user belongs to the experiment, then make the necessary changes to package.json.
+        // Once the API is final, we won't need to modify the package.json.
+        if (
+            this.serviceManager.get<IExperimentsManager>(IExperimentsManager).inExperiment(NativeNotebook.experiment) &&
+            this.useProposedApi
+        ) {
+            this.serviceManager.rebindSingleton<INotebookEditorProvider>(
+                INotebookEditorProvider,
+                NotebookEditorProvider
+            );
+        }
+
         const packageJsonFile = path.join(this.context.extensionPath, 'package.json');
         const content = JSON.parse(await this.fs.readFile(packageJsonFile));
