+++ src/client/datascience/notebook/serviceRegistry.ts
@@ -4,14 +4,10 @@
 'use strict';
 
 import { IExtensionSingleActivationService } from '../../activation/types';
-import { UseProposedApi } from '../../common/constants';
-import { NativeNotebook } from '../../common/experimentGroups';
-import { IExperimentsManager } from '../../common/types';
 import { IServiceManager } from '../../ioc/types';
-import { INotebookEditorProvider } from '../types';
 import { NotebookContentProvider } from './contentProvider';
 import { NotebookIntegration } from './integration';
-import { NotebookEditorProvider, NotebookEditorProviderActivation } from './notebookEditorProvider';
+import { NotebookEditorProviderActivation } from './notebookEditorProvider';
 import { NotebookKernel } from './notebookKernel';
 
 export function registerTypes(serviceManager: IServiceManager) {
@@ -25,13 +21,4 @@ export function registerTypes(serviceManager: IServiceManager) {
         IExtensionSingleActivationService,
         NotebookEditorProviderActivation
     );
-    // This condition is temporary.
-    // If user belongs to the experiment, then make the necessary changes to package.json.
-    // Once the API is final, we won't need to modify the package.json.
-    if (
-        serviceManager.get<IExperimentsManager>(IExperimentsManager).inExperiment(NativeNotebook.experiment) &&
-        serviceManager.get<boolean>(UseProposedApi)
-    ) {
-        serviceManager.rebindSingleton<INotebookEditorProvider>(INotebookEditorProvider, NotebookEditorProvider);
-    }
 }
